Add ref intent to string init/auto copy fns
Without 'ref' intent, these arguments will never be
wide when code generated. That means that these functions
could never be used on a remote string, which is an
oversight.

It's probably the case that we need to change the way that
the default intent works for records, since I'm not
sure that the current strategy of it always being
local makes sense.