Do not use calls to = in normalize for declared return types
If the return type is not declared, normalize generates code that uses
PRIM_MOVE to set the return value variable.

If the return type is declared however, before this patch, normalize
would generate a call to the = operator, starting with a
default-initialized left-hand side.

We would like declared return types to have the return value PRIM_MOVE'd
into the return variable. The issue is that if there is a declared return
type, we want casts to get added if the types are coercible, and an error
to be generated if not. That is really what = is doing in this case.

However, normalize cannot detect the types being the same or different
since the types are not known until function resolution.

This patch takes the approach of adding a new primitive, PRIM_COERCE,
that is added in normalize for such cases. PRIM_COERCE is handled by
function resolution. When the types are the same, the PRIM_COERCE is
simply removed. Otherwise, we check that = resolves (to generate errors
if a returned value is not coercible to the declared return type) and
then generate a cast call to perform the coercion.