Change PRIM_COERCE -> PRIM_COERCE_TO_RETURN
The version with PRIM_COERCE modified resolution in some ways that do not
fit with the rest of its design and implementation. In particular,
resolution tries to gather types from rhs to lhs (in move primitives),
but with PRIM_COERCE, the type information needed to flow from
lhs to rhs.

So, I changed PRIM_COERCE to PRIM_COERCE_TO_RETURN so that the type
of PRIM_COERCE is always known. As a result, I was able to provide
the function that returns its type so that it interacts naturally
with the rest of function resolution.

Other changes:
 - added a check for FLAG_TEMP before giving used-before-defined errors
   in normalize. I was getting errors with the return symbol not
   being defined. There is a comment indicating that this checking
   should perhaps not be done in normalize anyway.
 - updated normalize_returns to avoid using PRIM_NO_INIT ever.
   Instead, the AST generated always uses PRIM_MOVE to set the
   return value.
 - Minor adjustments to comments and removing changes that
   are no longer necessary.