Change to PRIM_COERCE with a type argument
Also, make fewer modifications of the FnSymbol->retType

My previous attempts were causing errors with recursive standalone
iterators. I realized that I could store the type more directly in the
primitive PRIM_COERCE_TO_RETURN and so this patch adds the type to coerce
to as the 2nd argument to the primitive instead of trying to make the
return type consistent or reflect anything in particular.

This approach is more flexible when the functions are transformed
multiple times in ways that change their return type (as is the case
especially with iterators).