Fix PRIM_COERCE handling for array promotion
In PR #2846, I updated the compiler to use PRIM_COERCE when returning
from a function with a declared return type. This patch adjusts the
changes in function resolution to fix a bug that was not revealed in
testing. In particular, a function like this:

  proc retit():[1..10] int {
    return 4;
  }

was causing compilation failures.

Another issue is that the code upgrading to = would use PRIM_NO_INIT
unconditionally, when for some types we should only use PRIM_INIT.
So this patch updates insertCasts to check propagateNotPOD to learn
if it is OK to use PRIM_NO_INIT.

Lastly, this patch adds a new test similar to the example shown
above that was incorrect but was apparently not revealed in testing.