Use initCopy instead of = when initializing a typed variable
Before this patch, normalize would use = to initialize r2 in
a case like this:
 var r2:R = r1;

but it would use initCopy to initialize r2 in this case:
 var r2 = r1;

As part of my campaign to remove the penalty for specifying types, and
building upon PR #2846, this patch adjusts normalize to use PRIM_COERCE
in this case just as we did for function returns.

I also added some TODO comments indicating that the common pattern like:
 var r2:R = someFunction();
or
 var r2 = someFunction();

should not generate an initCopy call for the result of someFunction
(since the return value is already the callee's).