Add PRIM_COERCE_INIT_COPY and use it
The previous approach of generating this kind of statement:
 var r1:R = r2;

as
 move r1, (coerce (initCopy r2), R)

Does not actually work for various reasons. One problem is that
in the case that the types do not match, we will use = on the
anyway and so the initCopy call would not be necessary.

To solve these issues, added a new primitive PRIM_COERCE_INIT_COPY
that behaves almost the same as PRIM_COERCE:

* if the types are not the same, init a temporary and call "="
* if the types are the same, instead of just using PRIM_MOVE,
  call initCopy on the right-hand-side.

Now the example:
 var r1:R = r2;

normalizes to
 move r1, (coerce_init_copy r2, R)