Disable infinite loop check for while statements
The compiler was attempting to raise an error for infinite loops but the
implementation is problematic:
 - it only applies to while statements
 - it does not handle 'return' or 'break'

Besides, it's not obvious to me that a program containing an infinite
loop is actually an error.

So, this commit disables that check for the near term.
See also issue #6532.

Updated deitz return tests for no infinite loop error