Draft of improving normalization of yields
Commit 8d3b2c4065d6466dbaadab0ae0c8444e6645dae6 introduced a separate
pass to create a different "yield temp" per yield, instead of having all
of the yields share one. In the commit message, it poitns out that it
would be better to adjust the normalizer.

Since an iterator can continue to run after the yield, memory management
(copy and destroy for records) rules are different for yields than for
returns. This change makes the AST better reflect this difference.